Refactor database configuration and update product price type to integer

+++ ticaretapp/schemas.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel
 class ProductBase(BaseModel):
     name: str
     description: str
-    price: float
+    price: int
 
 class ProductCreate(ProductBase):
     pass
